+++ config/dev.exs
@@ -9,6 +9,6 @@ config :ex_money,
   log_info: :info,
   log_success: :info,
   json_library: Jason,
-  exchange_rates_cache: :dets
+  exchange_rates_cache: Money.ExchangeRates.Cache.Dets
 
 config :ex_cldr, locales: ["en", "root", "zh-Hans", "zh"]